Lots of stuff (see below)
Refactored to remove the use of the Builder pattern, and instead simply
create a mutable abstraction. This changes the philosophy behind the
crate a bit; instead of trying to expose an interface that's arguably
more comfortable to work with for simpler cases, this exposes an
interface similar to that of the original AudioUnit's, while still being
safe and Rust'ic. The benefit here is that consumers can now use
features that AudioUnits provide out-of-the-box; for instance, switching
rendering callbacks and other properties on-the-fly. This allows them to
build more specific abstractions they need, intead of being confined by
our particular abstraction/usage pattern. The downside is that the usage
isn't quite as pretty, but safety is still guaranteed, errors are still
properly exposed, and we gain finer-grained control of the AudioUnit.
While the abstraction is still not complete, this also opens the door
for contributors to start abstracting other features of AudioUnits we
don't yet support in a very natural way, since the usage pattern is the
same as the underlying API, only safer.

The RenderCallback type has been changed from a function pointer
container to the function pointer type itself. This similifies many of
the abstraction's method signatures, which I think makes them easier to
read/understand.

Also replaced open/close functions with the more accurate start/stop
functions, that wrap the AudioUnit's own functions of the same names.

A few TODO's were added in the process; going to go over those next and
perhaps also expose more of the original AudioUnit interface.