Regenerate Github Auth Token
Ivan Vyshnevskyi kindly took the time to point out that the auth token
for this repo has been leaked.

This PR adds a regenerated token along with an adjustment to the
.travis.yml that should avoid future leaks.

This is the same issue as PistonDevelopers/image#596.

Ivan will be releasing a blog post on how he discovered the issue soon.
Will link to it as soon as it is available!